Working on adding parser support for initialization blocks

+++ j--/src/jminusminus/JLiteralFloat.java
@@ -37,7 +37,7 @@ class JLiteralFloat extends JExpression {
      */
 
     public JExpression analyze(Context context) {
-        type = Type.INT;
+        type = Type.DOUBLE;
         return this;
     }
